Use one version of file s-bbsumu.adb
When building libgnat via runtime_build.gpr, the file s-soflin.adb indirectly
references s-bbsumu.adb, but because runtime_build.gpr does not know that
we've renamed that file (to e.g. s-bbsumu-2.adb) since that is set in
ravenscar_build.gpr for libgnarl, it can;t find s-bbsumu.adb and errors out.

To resolve this, we instead use one version of s-bbsumu.adb with guards to
check whether it's a multiprocessing runtime to stub out the public subprograms
for single-processor builds.